+++ lib/WorseReflection/Core/Inference/Resolver/StringLiteralResolver.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Phpactor\WorseReflection\Core\Inference\Resolver;
+
+use Microsoft\PhpParser\Node;
+use Microsoft\PhpParser\Node\StringLiteral;
+use Phpactor\WorseReflection\Core\Inference\Frame;
+use Phpactor\WorseReflection\Core\Inference\NodeContext;
+use Phpactor\WorseReflection\Core\Inference\NodeContextFactory;
+use Phpactor\WorseReflection\Core\Inference\Resolver;
+use Phpactor\WorseReflection\Core\Inference\Symbol;
+use Phpactor\WorseReflection\Core\Inference\SymbolContextResolver;
+use Phpactor\WorseReflection\Core\TypeFactory;
+use Phpactor\WorseReflection\Core\Util\NodeUtil;
+
+class StringLiteralResolver implements Resolver
+{
+    public function resolve(SymbolContextResolver $resolver, Frame $frame, Node $node): NodeContext
+    {
+        assert($node instanceof StringLiteral);
+        return NodeContextFactory::create(
+            (string) $node->getStringContentsText(),
+            $node->getStartPosition(),
+            $node->getEndPosition(),
+            [
+                'symbol_type' => Symbol::STRING,
+                'type' => TypeFactory::string(),
+                'value' => (string) $node->getStringContentsText(),
+                'container_type' => NodeUtil::nodeContainerClassLikeType($resolver->reflector(), $node),
+            ]
+        );
+    }
+}
